--- fuzzy/fuzz_partition_test.go
@@ -1,10 +1,0 @@
-package fuzzy
-
-/*
-func TestAA(t *testing.T) {
-	jitterMax := 1 * time.Second
-	for i := 0; i < 10; i++ {
-		fmt.Println(time.Duration(uint64(rand.Int63()) % uint64(jitterMax)))
-	}
-}
-*/
